Move computer methods out of Gameplay and into their own Computer Class and Test files
